requirements.txt
streamlit==1.28.1
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
requests==2.31.0
Pillow==10.0.1
streamlit-option-menu==0.3.6
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1 